[MRG] LogisticRegression: Check if l1_ratio is None when using ElasticNet
LogisticRegression with `penalty=elasticnet` requires the user specify the `l1_ratio` between the L1 and L2 penalties, but the code currently does not check for this. 

Currently, an unhelpful error stating "unsupported operand type(s) for -: 'int' and 'NoneType'" is raised when `fold_coefs_` is calculated later in `fit()`.

With the proposed change, it will be clear to the user that the error was not specifying `l1_ratio`.