DOC GPs: log_marginal_likelihood() log(theta) input
Document the fact that the input argument `theta` to
GaussianProcess{Regressor,Classifier}'s `log_marginal_likelihood()`
method is actually expected to be `log(theta)`, as this is used in the
optimizer code path when `self.kernel_.theta` is passed and where the
latter triggers a getter that returns `log(theta)`. This is now
consistent with other doc strings where "log-transformed theta" is
mentioned.

In the method doc strings, only add "log-transformed". In the GPR user
guide add an extra note.