Added precision_recall_fscore_support_pred and output_pred.
A new function called precision_recall_fscore_pred was added
that outputs the number of predictions along with support.
This new function was added because modifying the output of an
existing function didn't seem like a good idea.

output_pred is a kwarg that is Boolean and tells the classification_
report function whether or not to output the number of predictions
as well._