MAINT: remove `from numpy.math cimport` statements
All C99 constants and functions are available from Cython's `libc`,
which is cleaner. NumPy would like to get rid of shipping
libnpymath, and has also deprecated the (unrelated) `numpy.math`
in the Python API. Finally, numpy is shipping its own `__init__.pxd`
now, but not `math.pxd` - so that should be considered legacy at
this point. So best to not use it at all.