[ENH] Adding leaf node samples to be stored when "quantile" tree is turned on (#45)
#### Reference Issues/PRs
Addresses the quantile-trees part of:
https://github.com/neurodata/scikit-tree/issues/29


#### What does this implement/fix? Explain your changes.
1. Stores for each leaf node a 2D numpy array of the y-samples (remember
`y` is (n_samples, n_outputs))
2. Does this all the way in Criterion
3. Only supports supervised tree/splitter/criterion
4. merges in `main` changes.

#### Any other comments?


<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:

http://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->

---------

Signed-off-by: Adam Li <adam2392@gmail.com>