DOC fix description of tolerance in OMP
The phrasing of the documentation suggested that the result was so that the resulting regression coefficients had the property that the norm of the residual was less than `tol`, when in fact all that was guaranteed was that the norm square would be less than `tol`. Looking at the implementation, the latter is very much the intended behavior, so we change the documentation instead of the implementation.