MAINT: Adapt sklearn for NumPy default integer change
This adepts the `_random.pyx` file to return whatever is the NumPy
default integer, which, on NumPy 2.0 would fix.
Since the cython symbol wasn't used, I just removed it as it bites
with the overloading.

See https://github.com/numpy/numpy/pull/24224 for the commit which
would make this necessary.

At the time this is a bit hard to test since the SciPy nightlies are
incompatible with that NumPy branch.  But I thought I would put it out
there for discussion.

The alternative and simpler solution might be to just force 64bit
results on any 64bit system and not worry about the NumPy version.