Update getting_started.rst
The existing documentation uses accuracy_score which can't be used on continuous data. I switched to r2_score. I am not an expert, so unclear if use of regression metric is good enough or if there should be some modification to the predicted output to keep using integer scoring.

```
# pipe.predict(X_test) returns continuous results but accuracy_score can't handle that directly
ValueError                                Traceback (most recent call last)
Cell In[36], line 1
----> 1 accuracy_score(pipe.predict(X_test), y_test)
...
ValueError: Classification metrics can't handle a mix of continuous and multiclass targets
```