CI: Build nightly wheels using NumPy 2 pre-releases (and others)
NumPy 2.0 needs building with NumPy 2.0 (very soon) to allow some
larger changes (technically, this is already necessary).

Thus, nightlies must be build with NumPy 2.0 (it is better to not
test them with it, which this does).

This takes the simplified route of just using any pre-releases during
nightly builds.  If that fails (say cython is broken for a bit) then
this could be adapted to explicitly install numpy and using
`--no-build-isolation` instead.
(Might require a bit of care to not use the same env when testing, not sure)