GroupKFold: Use stable sort to produce consistent splits across installs
The use of np.argsort with unstable sorting results in different splits
across installations.  This one is hard to write a test for since the
splits will always be consistent with multiple runs in the same
environment.  Ideas of how to test this would be appreciated, but
perhaps it's a simple enough change to motivate without a failing test.