GCC 14: Remove `inline` for `rdist_csr` and `rdist`
This is related with cython/cython#2747

This problem was mentioned in a bug open to keep track of the effort to remove all compilation warnings #24875 (search "Incompatible pointers types casts")

With gcc < 14 the code compiles with incompatible-pointer warning. But with gcc 14 that warning is an error.

Fedora bug: https://bugzilla.redhat.com/show_bug.cgi?id=2261602

A posible workaround is to remove the inline in the definition of the rdist_csr and rdistmembers.

With that change, cython generates correct code (I don't know why). But removing the inline can impact the performance.

Fixes https://github.com/scikit-learn/scikit-learn/issues/28530

Signed-off-by: Wong Hoi Sing Edison <hswong3i@pantarei-design.com>