MAINT: convert `numpy.array_api` to array-api-strict
The `numpy.array_api` module has been converted into a standalone
package (`array-api-strict`). This new package is stable and has had a
1.0 release. The `numpy.array_api` module was marked experimental, and
will be removed for NumPy 2.0. Since `array-api-strict` works with both
NumPy 1.2x and NumPy 2.0, using that for testing compliance to the
syntax and semantics of the array API standard should always be
preferred over still trying to use `numpy.array_api`.