MAINT Fix builder partial fit (#62)
#### Reference Issues/PRs
Fixes state of builder_ to not need to be maintained.

Prolly needs unit-tests to determine if this "functions as desired".
I.e.
- changing datatype of X over multiple partial fits should fail nicely, 
- changing datatype of y
- classification and regression

#### What does this implement/fix? Explain your changes.


#### Any other comments?


<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:

https://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->

---------

Signed-off-by: Adam Li <adam2392@gmail.com>