Fix: handling decision functions wrongly interpreted as probabilities.
It's possible that the values of the decision functions, ranging from 0 to 1, were misinterpreted as probabilities when they weren't. By introducing a new parameter to `_ovr_decision_function` containing the indexes of the classifiers for which confidences are predicted probabilities, we can now correctly handle cases where confidences are decision functions or predicted probabilities for the positive class for each binary classifier.