RFE and RFECV allow features_to_select to be larger than available features (scikit-learn#28725)
Implemented a warning message for when the RFE or the RFECV objects are initialized with a n_features_to_select or min_features_to_select (respectively) attribute larger than the number of features present in the X variable that is passed to the fit method.

Additionally, a new test function has been added to ensure the correct warning message is raised in these scenarios. The test comprehensively covers both cases and validates that the warning is emitted as expected.