Issue #29783 solved
Issue #29783 Solved:
Added Check for coef_: Using hasattr to check if coef_ exists on the estimator_ attribute of RFECV before accessing it. This prevents the AttributeError and ensures that the code only attempts to access coef_ if it is available.
And code became robust after giving try and except condition