WIP, MAINT: NeighborsBase `KDTree` upstream
* This is an extension of the concept in gh-31347--here,
part of the usage of in-house `KDTree` in `NeighborsBase`
is replaced by its upstream version from SciPy. This is a much
more challenging effort that clearly shows some substantial
differences between the two `KDTree` APIs/methods, and the shims
needed to address them. At the moment, there is still a small
number of residual test failures (29 locally) in the full testsuite.

* Some kind of API unification/equivalence of offerings seems likely
to be needed for these kinds of replacements to be more sustainable (the
shims added here were quite time consuming to figure out). Some of the
test expectations may also be debatable for cases with i.e., degenerate
input.